requirements.txt
langchain==0.3.21
langchain-community==0.3.20
langchain-openai==0.3.9
chromadb==0.5.23
sentence-transformers==3.4.1
python-dotenv==1.0.1
pydantic==2.10.6
